*: move "discovery" to "internal/discovery"
