+++ etcdserver/server.go
@@ -30,7 +30,6 @@ import (
 	"time"
 
 	"github.com/coreos/etcd/auth"
-	"github.com/coreos/etcd/discovery"
 	"github.com/coreos/etcd/etcdserver/api"
 	"github.com/coreos/etcd/etcdserver/api/v2http/httptypes"
 	pb "github.com/coreos/etcd/etcdserver/etcdserverpb"
@@ -38,6 +37,7 @@ import (
 	"github.com/coreos/etcd/etcdserver/stats"
 	"github.com/coreos/etcd/internal/alarm"
 	"github.com/coreos/etcd/internal/compactor"
+	"github.com/coreos/etcd/internal/discovery"
 	"github.com/coreos/etcd/internal/raftsnap"
 	"github.com/coreos/etcd/lease"
 	"github.com/coreos/etcd/lease/leasehttp"
