+++ internal/discovery/discovery_test.go
